feat(activities): isolate backend controlers
