+++ src/app/(dashboard)/impact/page.tsx
@@ -1,5 +1,4 @@
 import ImpactLog from '@/components/impact/ImpactLog'
-import StatsImpact from '@/components/impact/StatsImpact'
 import { Stack } from '@mui/material'
 import { type Metadata } from 'next'
